* use shared build layer script in all functions
* fix layers permission issue on linux

+++ sources/build_custom_layer.sh
@@ -1,12 +1,21 @@
-rm -rf gems
+#!/bin/sh
 
-docker run --rm -v $PWD:/var/layer \
-           -w /var/layer \
-           amazon/aws-sam-cli-build-image-ruby2.7 \
-           bundle install --path=gems
+image=amazon/aws-sam-cli-build-image-ruby2.7
+module=$1
 
-mv gems/ruby/* gems/ \
+if [ -z "$module" ]; then
+  echo "Usage: $0 <module>"
+  exit 1
+fi
+
+docker run --rm -v $PWD/$module:/var/layer -w /var/layer $image sh -c "
+  set -x \
+  && rm -rf gems \
+  && bundle config set --local path 'gems' \
+  && bundle install \
+  && mv gems/ruby/* gems/ \
   && rm -rf gems/2.7.0/cache \
-  && rm -rf gems/ruby
-
-rm -rf .bundle
+  && rm -rf gems/ruby \
+  && rm -rf .bundle \
+  && chown -R $(id -u):$(id -g) gems/
+"
